+++ js/data.js
@@ -1,0 +1,49 @@
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner'];
+
+const TITLES = [
+  'Большая уютная квартира',
+  'Маленькая неуютная квартира',
+  'Огромный прекрасный дворец',
+  'Маленький ужасный дворец',
+  'Красивый гостевой домик',
+  'Некрасивый негостеприимный домик',
+  'Уютное бунгало далеко от моря',
+  'Неуютное бунгало по колено в воде',
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+
+const HOURS = ['12:00', '13:00', '14:00'];
+
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+const DESCRIPTIONS = [
+  'Невероятный дворец с аквадискотекой и театром',
+  'Светлая и просторная квартира с видом на набережную',
+  'Небольшой загородный домик в скандинавском стиле',
+  'Уютная студия в центре города со всем необходимым',
+];
+
+export {
+  FEATURES,
+  TITLES,
+  TYPES,
+  HOURS,
+  PHOTOS,
+  DESCRIPTIONS
+};
